requirements.txt
typhoon-ocr
python-dotenv
PyMuPDF
tqdm
colorama
